add : enabled unlimited directory

+++ src/lib-header/fat32.h
@@ -246,7 +246,7 @@ int8_t write(struct FAT32DriverRequest request);
  */
 int8_t delete(struct FAT32DriverRequest request);
 
-int8_t dirtable_linear_search(uint32_t parent_cluster_number, struct FAT32DriverRequest entry);
+int32_t dirtable_linear_search(uint32_t parent_cluster_number, struct FAT32DriverRequest entry);
 bool check_dir_valid(uint32_t parent_cluster_number);
 bool is_file_exists(uint32_t parent_cluster_number, struct FAT32DriverRequest entry);
 void add_to_dir_table(uint32_t parent_cluster_number, struct FAT32DriverRequest entry,int16_t entry_cluster);
